// src/app/models/clps.ts
export class CLPS {
    id?: string = '';
    shipId?: string = '';
    fty?: string = '';
    po?: string = '';
    itemNo?: string = '';
    rrd?: Date;
    erd?: Date;
    notes?: string = '';
    dateCreated?: string = '';
    cbmCtn?: number = 0;
    qty?: number = 0;
    cost?: number = 0;
    extCost?: number = 0;
    cbm?: number = 0;
    gwCtn?: number = 0;
    gw?: number = 0;
    etd?: Date;
    cntrNo?: number = 0;
    webId?: string = '';
    orgPoQty?: number = 0;
    clp?: string = '';
    poDate?: Date;
    loc?: string = '';
    shipVia?: string = '';
    poLine?: string = '';
    splitQty?: number;
    newSplitVal?: number;
    newEtd?: string = '';

    newEta?: string = '';
    list_id?: string;
    erdDt?: string
    rrdDt?: string
    poDt?: string
    isParent?: boolean = false;
    splitFrom?: string = '';

    bk?: string = '';
    cCost?: number = 0;
    chgbk?: string = '';
    consignee?: string = '';
    cont?: string = '';
    cube?: number = 0;
    do?: string = '';
    extCube?: number = 0
    fdaDesc?: string = '';
    gri?: string = '';
    h?: number = 0;
    htsCode?: string = '';
    htsValue?: number = 0;
    l?: number = 0;
    lb?: number = 0;
    logo?: string = ''
    pcscs?: string = ''
    tariffCode?: string = '';
    tariffValue?: number = 0;
    terminal?: string = ''
    usc?: string = '';
    w?: number = 0;
    eta?: Date;
    userOperation?: string = '';
    dirName?: string;
    uniKey?: string;
    isVisible?: boolean = false;
    isSplit?: boolean = false

}